# Conflicts:
#	.idea/inspectionProfiles/Project_Default.xml
#	src/main/java/org/example/Main.java
#	src/main/java/org/example/controller/StudentController.java
#	src/main/java/org/example/controller/admin/AdminController.java
#	src/main/java/org/example/entity/BoardArticle.java
#	src/main/java/org/example/entity/BoardInfo.java
#	src/main/java/org/example/entity/GradeInfo.java
#	src/main/java/org/example/entity/LectInfo.java
#	src/main/java/org/example/entity/LectNth.java
#	src/main/java/org/example/entity/LmsConts.java
#	src/main/java/org/example/entity/Member.java
#	src/main/java/org/example/repository/GradeInfoRepository.java
#	src/main/java/org/example/repository/LectNthRepository.java
#	src/main/java/org/example/repository/LmsContsRepository.java
#	src/main/java/org/example/repository/admin/AdminLectInfoRepository.java
#	src/main/java/org/example/repository/admin/AdminRepository.java
#	src/main/java/org/example/repository/admin/AdminStudLectProgRepository.java
#	src/main/resources/templates/prof/assiGrade.html

+++ src/main/java/org/example/dto/CheckSemGradeDto.java
@@ -13,7 +13,7 @@ public class CheckSemGradeDto {
     private String semYear;
     private String semSem;
     private Long semCrecpl;
-    private Long semAvgCrecpl;
+    private double semAvgCrecpl;
 //    private String grade;
 //    private Long lectCredit;
